% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-fetching-data.R
\name{doi.fetch_metadata}
\alias{doi.fetch_metadata}
\title{Fetching metadata for a doi from COCI}
\usage{
doi.fetch_metadata(DOI)
}
\arguments{
\item{DOI}{digital object identifier as character}
}
\description{
Fetches reference, citation and meta information from the OpenCitations
COCI project, see here: https://opencitations.net/index/coci
}
