% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-file-management.R
\name{local.paper.exists}
\alias{local.paper.exists}
\title{Checks whether local paper .pdf exists}
\usage{
local.paper.exists(DOI)
}
\arguments{
\item{DOI}{digital object identifier as character}
}
\description{
Checks whether local paper .pdf exists
}
