% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-file-management.R
\name{local.paper_data.exists}
\alias{local.paper_data.exists}
\title{Checks whether local paper data exists}
\usage{
local.paper_data.exists(DOI)
}
\arguments{
\item{DOI}{digital object identifier as character}
}
\description{
Checks whether local paper data exists
}
