% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-file-management.R
\name{local.paper_data_path}
\alias{local.paper_data_path}
\title{Returns path to paper_data.yaml file}
\usage{
local.paper_data_path(DOI)
}
\arguments{
\item{DOI}{digital object identifier as character}
}
\description{
Returns path to paper_data.yaml file
}
