% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-file-management.R
\name{local.paper_dir}
\alias{local.paper_dir}
\title{Returns the directory of a paper}
\usage{
local.paper_dir(DOI)
}
\arguments{
\item{DOI}{digital object identifier}
}
\description{
Returns the directory of a paper
}
