% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-file-management.R
\name{local.paper_path}
\alias{local.paper_path}
\title{Returns the path to a paper .pdf}
\usage{
local.paper_path(DOI, filename = NULL)
}
\arguments{
\item{DOI}{digital object identifier as character}

\item{filename}{filename to save as as character eg. paper_name.pdf}
}
\description{
Returns the path to a paper .pdf
}
