% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-file-management.R
\name{local.papers.downloaded}
\alias{local.papers.downloaded}
\title{Returns data for the papers you've downloaded}
\usage{
local.papers.downloaded()
}
\description{
Returns data for the papers you've downloaded
}
