% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-data-accessors.R
\name{paper.data}
\alias{paper.data}
\title{Returns the yaml data from a local paper}
\usage{
paper.data(DOI)
}
\arguments{
\item{DOI}{digital object identifier as character}
}
\description{
Returns the yaml data from a local paper
}
