% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-downloading.R
\name{paper.download}
\alias{paper.download}
\title{Downloads a paper}
\usage{
paper.download(
  DOI,
  download_path = local.paper_path(DOI),
  iframe_xpath = "//*[@id='pdf']"
)
}
\arguments{
\item{DOI}{digital object identifier as character}

\item{download_path}{full path to download .pdf to as a character}

\item{iframe_xpath}{xpath attribute to find the .pdf in, as a character with escapes if necessary}
}
\description{
This requires a the INSTITUTION_PDF_PREFIX env variable set according to paper.institution_doi_url
Visits a institution/DOI url, searches for an iframe object with an embedded pdf according to a
given xpath, and then downloads the paper to a location set with download_path
}
