% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-downloading.R
\name{paper.download_and_open}
\alias{paper.download_and_open}
\title{Downloads and opens a paper}
\usage{
paper.download_and_open(DOI, download_path = local.paper_path(DOI))
}
\arguments{
\item{DOI}{digital object identifier as character}

\item{download_path}{full path to download .pdf to as a character}
}
\description{
See paper.download function for details
}
