% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-fetching-data.R
\name{paper.fetch_citations_data}
\alias{paper.fetch_citations_data}
\title{Fetches the citations/references/meta for all works citing a DOI}
\usage{
paper.fetch_citations_data(DOI, download_paper = FALSE)
}
\arguments{
\item{DOI}{digital object identifier as a string}

\item{download_paper}{whether to download the paper .pdf as well}
}
\description{
Fetches the citations/references/meta for all works citing a DOI
}
