% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-fetching-data.R
\name{paper.fetch_data}
\alias{paper.fetch_data}
\title{Get metadata for a paper}
\usage{
paper.fetch_data(
  DOI,
  download_paper = TRUE,
  auto_open_paper = TRUE,
  progress = NULL
)
}
\arguments{
\item{DOI}{digital object identifier as a character}

\item{download_paper}{whether or not to include the paper .pdf as well}

\item{auto_open_paper}{whether to open the paper as soon as it's downloaded, if download_paper is true}

\item{progress}{a progress bar object, to be ticked after completion}
}
\description{
Get metadata for a paper
}
