% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-fetching-data.R
\name{paper.fetch_references_data}
\alias{paper.fetch_references_data}
\title{Fetches the citations/references/meta for all references in a DOI}
\usage{
paper.fetch_references_data(DOI, download_paper = FALSE)
}
\arguments{
\item{DOI}{digital object identifier as character}

\item{download_paper}{logical, whether to download the .pdf too}
}
\description{
Fetches the citations/references/meta for all references in a DOI
}
