% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-data-accessors.R
\name{paper.find}
\alias{paper.find}
\title{Find paper in local directory by name}
\usage{
paper.find(value, by = "name", only_downloaded = TRUE)
}
\arguments{
\item{value}{name to search for}

\item{by}{currently can only search by name}

\item{only_downloaded}{currently can only search directories with a downloaded .pdf}
}
\description{
Find paper in local directory by name
}
