% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-downloading.R
\name{paper.institution_doi_url}
\alias{paper.institution_doi_url}
\title{Returns a URL for accessing a pdf}
\usage{
paper.institution_doi_url(DOI)
}
\arguments{
\item{DOI}{Digital Object Identifier as a character}
}
\description{
If your institution has a forwarding URL which allows direct access to .pdf
papers, by appending a URL to the paper DOI, set that URL to the environmental
variable INSTITUTION_PDF_PREFIX and this method will produce links to .pdfs
from DOIs
}
